requirements.txt
azure-functions
python-decouple
sqlalchemy
sendgrid